[#169167769] Added temporarily blocked card screen
Refactoring.
