+++ locales/en/cie/temporary_blocked_cie.md
@@ -1,0 +1,6 @@
+# Card temporarily blocked
+
+The card used was blocked because you tried to log in too many times with an incorrect PIN.  
+No problem: you can unlock it with the **PUK code** which was delivered to you along with the card.  
+To complete this operation you must however **use a special free app**, which is called **CieID**.  
+You will be able to log in to IO again when you unlock the card.
