+++ locales/it/cie/temporary_blocked_cie.md
@@ -1,0 +1,6 @@
+# Carta bloccata temporaneamente
+
+La carta utilizzata è stata bloccata perché hai provato ad accedere troppe volte con un PIN errato.  
+Non c'è problema: puoi sbloccarla con il **codice PUK** che ti è stato consegnato insieme alla carta.  
+Per completare questa operazione devi però **utilizzare un'apposita app gratuita**, che si chiama **CieID**.  
+Potrai accedere nuovamente ad IO quando avrai sbloccato la carta.
